requirements.txt
absl-py==0.10.0
aiodns==2.0.0
aiohttp==3.6.2
aiohttp-socks==0.3.9
aniso8601==8.1.0
app==0.0.1
appdirs==1.4.4
argon2-cffi==20.1.0
astor==0.8.1
astroid==2.0.4
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==19.3.0
audioread==2.1.9
backcall==0.2.0
backports.weakref==1.0rc1
beautifulsoup4==4.9.0
bleach==1.5.0
Box2D==2.3.10
cachetools==4.1.1
cchardet==2.1.6
certifi==2020.4.5.1
cffi==1.11.5
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.0
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
discord.py==0.16.0
elasticsearch==7.7.0
entrypoints==0.3
fake-useragent==0.1.11
fire==0.1.3
Flask==1.1.2
Flask-Ext==0.1
Flask-MySQL==1.5.1
Flask-RESTful==0.3.8
flatbuffers==1.12
future==0.17.1
gast==0.3.3
geographiclib==1.50
geopy==1.22.0
google-auth==1.21.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
gym==0.18.0
h5py==2.10.0
html5lib==0.9999999
idna==2.8
idna-ssl==1.1.0
imageio==2.9.0
importlib-metadata==2.0.0
ipykernel==5.4.2
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.4
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-rl==0.4.2
kiwisolver==1.3.1
lazy-object-proxy==1.3.1
librosa==0.8.0
llvmlite==0.34.0
lxml==4.5.2
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
mistune==0.8.4
multidict==4.7.5
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.2
nltk==3.5
notebook==6.1.5
numba==0.51.2
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.4.0.46
opt-einsum==3.3.0
packaging==20.4
pandas==1.0.3
pandocfilters==1.4.3
parso==0.7.1
pickleshare==0.7.5
Pillow==7.2.0
pooch==1.2.0
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.15.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==3.1.1
pycparser==2.19
pyglet==1.5.0
Pygments==2.7.2
pylint==2.1.1
PyMySQL==0.10.1
PyNaCl==1.0.1
pyparsing==2.4.7
PyQt5==5.15.0
PyQt5-sip==12.8.0
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
pywin32==228
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==19.0.2
qtconsole==5.0.1
QtPy==1.9.0
rebulk==2.0.1
regex==2017.4.5
requests==2.21.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.6
schedule==0.6.0
scikit-learn==0.23.2
scipy==1.4.1
sounddevice==0.4.1
SoundFile==0.10.3.post1
soupsieve==2.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
textblob==0.15.3
threadpoolctl==2.1.0
toposort==1.5
torch==1.8.0
torchvision==0.9.0
tornado==6.1
tqdm==4.31.1
traitlets==4.3.3
tweepy==3.9.0
twint==2.1.4
typed-ast==1.1.0
typing==3.7.4.1
typing-extensions==3.7.4.2
urllib3==1.24.3
wcwidth==0.2.5
websockets==3.4
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yarl==1.4.2
zipp==3.2.0
